64544 - Add onboarding. P2 ViewModel
