+++ Sveti/Support/Reusable UIComponents/Onboarding/Models/Interfaces/IOnboardingController.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+protocol IOnboardingController {
+  var viewModel: IOnboardingVM { get set }
+  func presentIfNeeded()
+  func move(to: OnboardingMoveDirection)
+}
