+++ Sveti/Support/Reusable UIComponents/Onboarding/Models/Interfaces/IOnboardingVM.swift
@@ -1,12 +1,9 @@
 import Foundation
 
-protocol IOnboardingController {
+protocol IOnboardingVM {
   var onboardingKey: String { get set }
   func updateOnboardingWatchStatus()
   func getOnboardingWatchStatus() -> Bool
 
   var slides: [OnboardingSlide] { get set }
-  func move(to: OnboardingMoveDirection)
-
-  func presentIfNeeded()
 }
