+++ Sveti/Support/Reusable UIComponents/Onboarding/Views/OnboardingVC.swift
@@ -2,26 +2,16 @@ import UIKit
 
 class OnboardingVC: UIViewController, IOnboardingController {
 
-  var onboardingKey: String
-  var slides: [OnboardingSlide]
+  var viewModel: IOnboardingVM
 
-  init(with slides: [OnboardingSlide], key: String) {
-    self.onboardingKey = key
-    self.slides = slides
-
+  init(viewModel: IOnboardingVM) {
+    self.viewModel = viewModel
     super.init(nibName: nil, bundle: nil)
+
   }
 
   required init?(coder: NSCoder) {
     fatalError("init(coder:) has not been implemented")
-  }
-
-  func updateOnboardingWatchStatus() {
-    //
-  }
-
-  func getOnboardingWatchStatus() -> Bool {
-    return false
   }
 
   func move(to: OnboardingMoveDirection) {
