+++ Sveti/Support/Reusable UIComponents/Onboarding/Views/OnboardingVM.swift
@@ -1,0 +1,23 @@
+import Foundation
+
+class OnboardingVM: IOnboardingVM {
+
+  var onboardingKey: String
+  var slides: [OnboardingSlide]
+
+  init(key: String, slides: [OnboardingSlide]) {
+    self.onboardingKey = key
+    self.slides = slides
+  }
+
+  func updateOnboardingWatchStatus() {
+    //
+  }
+
+  func getOnboardingWatchStatus() -> Bool {
+    true
+  }
+
+}
+
+
